+++ docs/changes.rst
@@ -29,6 +29,7 @@ Weblate 5.16
 * Adding plural strings with singular matching existing string is now prohibited for bilingual translations (see :ref:`bimono`).
 * Automatic :ref:`component-repoweb` URL for common code hosting sites.
 * File formats that only differed in encoding (CSV, GWT Properties, Java Properties, iOS Strings) have been merged into single formats.
+* Gracefully handle invalid check flags in :ref:`custom-checks`.
 
 .. rubric:: Compatibility